refactor : 회원 상세 정보 조회 방법 변경
