+++ src/main/java/com/study/shoestrade/repository/interest/InterestProductRepositoryCustom.java
@@ -12,6 +12,4 @@ public interface InterestProductRepositoryCustom {
     List<InterestProduct> findPreInterests(String email, Long productId);
 
     Page<MyInterest> findMyInterests(String email, Pageable pageable);
-
-    List<MyInterest> findMemberInterests(Long id);
 }
